consolidate mode switching in geoportal and fix cursor style handling conflicts
