+++ apps/geoportal/src/app/components/GeoportalMap/controls/MapWrapper.tsx
@@ -84,7 +84,6 @@ import {
   getUIMode,
   getZenMode,
   setZenMode,
-  toggleUIMode,
   UIMode,
 } from "../../../store/slices/ui.ts";
