+++ apps/geoportal/src/app/hooks/useMapModes.ts
@@ -2,7 +2,6 @@ import { useEffect, useRef } from "react";
 import { useDispatch, useSelector } from "react-redux";
 import { useMapMeasurementsContext } from "@carma-commons/measurements";
 import { getUIMode, setUIMode, toggleUIMode, UIMode } from "../store/slices/ui";
-import { useFeatureInfoModeCursorStyle } from "./useFeatureInfoModeCursorStyle";
 import { cancelOngoingRequests } from "../components/GeoportalMap/topicmap.utils";
 
 export const useMapModes = () => {
@@ -11,8 +10,6 @@ export const useMapModes = () => {
   const { isMeasurementEnabled, setMeasurementEnabled } =
     useMapMeasurementsContext();
   const prevUiModeRef = useRef(uiMode);
-
-  useFeatureInfoModeCursorStyle();
 
   useEffect(() => {
     const justSwitchedToFeatureInfo =
@@ -33,17 +30,6 @@ export const useMapModes = () => {
     }
     prevUiModeRef.current = uiMode;
   }, [isMeasurementEnabled, uiMode, dispatch, setMeasurementEnabled]);
-
-  // Prevent cursor race condition when switching from Measurement to Feature Info
-  // The measurement tool resets cursor to 'pointer' when disabled, overriding Feature Info's 'crosshair'
-  useEffect(() => {
-    if (uiMode === UIMode.FEATURE_INFO && !isMeasurementEnabled) {
-      const mapElement = document.getElementById("routedMap");
-      if (mapElement) {
-        mapElement.style.cursor = "crosshair";
-      }
-    }
-  }, [uiMode, isMeasurementEnabled]);
 
   const handleToggleFeatureInfo = () => {
     cancelOngoingRequests();
